feat!: update some integrations packaging/bundling
