+++ examples/react-router/plugin/package.json
@@ -11,7 +11,6 @@
   },
   "dependencies": {
     "@generouted/react-router": "workspace:*",
-    "generouted": "workspace:*",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-router-dom": "6.9.0"
