+++ examples/react-router/plugin/src/main.tsx
@@ -1,5 +1,5 @@
 import { createRoot } from 'react-dom/client'
-import { Routes } from 'generouted/react-router'
+import { Routes } from '@generouted/react-router'
 
 const container = document.getElementById('app')!
 createRoot(container).render(<Routes />)
