+++ examples/react-router/plugin/src/pages/_app.tsx
@@ -1,5 +1,5 @@
 import { Outlet } from 'react-router-dom'
-import { Modals } from 'generouted/react-router'
+import { Modals } from '@generouted/react-router'
 
 import { Link, useModals, useNavigate, useParams } from '../router'
