+++ examples/react-router/plugin/vite.config.ts
@@ -1,6 +1,6 @@
 import { defineConfig } from 'vite'
 import react from '@vitejs/plugin-react'
-import generouted from '@generouted/react-router'
+import generouted from '@generouted/react-router/plugin'
 
 export default defineConfig({
   plugins: [react(), generouted()],
