+++ examples/solid-router/package.json
@@ -11,7 +11,6 @@
   "dependencies": {
     "@generouted/solid-router": "workspace:*",
     "@solidjs/router": "^0.8.2",
-    "generouted": "workspace:*",
     "solid-js": "^1.6.15"
   },
   "devDependencies": {
