+++ examples/solid-router/src/main.tsx
@@ -1,4 +1,4 @@
 import { render } from 'solid-js/web'
-import { Routes } from 'generouted/solid-router'
+import { Routes } from '@generouted/solid-router'
 
 render(Routes, document.getElementById('app') as HTMLElement)
