+++ examples/solid-router/vite.config.ts
@@ -1,6 +1,6 @@
 import { defineConfig } from 'vite'
 import solid from 'vite-plugin-solid'
-import generouted from '@generouted/solid-router'
+import generouted from '@generouted/solid-router/plugin'
 
 export default defineConfig({
   plugins: [solid(), generouted()],
