+++ plugins/react-router/src/index-lazy.ts
@@ -1,0 +1,1 @@
+export * from 'generouted/react-router-lazy'
