+++ plugins/react-router/src/plugin/generate.ts
