+++ plugins/react-router/src/plugin/options.ts
