+++ plugins/react-router/src/plugin/template.ts
