+++ plugins/react-router/src/react.js
@@ -1,0 +1,2 @@
+import React from 'react'
+export { React }
