+++ plugins/solid-router/readme.md
@@ -11,11 +11,8 @@ In case you don't have a Vite project with Solid and TypeScript, check out this 
 ### Installation
 
 ```shell
-pnpm add @generouted/solid-router generouted @solidjs/router
+pnpm add @generouted/solid-router @solidjs/router
 ```
-
-- `generouted` provides the file-based routes
-- `@generouted/solid-router` generates types and type-safe router component/hooks
 
 ### Setup
 
@@ -24,7 +21,7 @@ pnpm add @generouted/solid-router generouted @solidjs/router
 
 import { defineConfig } from 'vite'
 import solid from 'vite-plugin-solid'
-import generouted from '@generouted/solid-router'
+import generouted from '@generouted/solid-router/plugin'
 
 export default defineConfig({ plugins: [solid(), generouted()] })
 ```
@@ -35,7 +32,7 @@ export default defineConfig({ plugins: [solid(), generouted()] })
 // src/main.tsx
 
 import { render } from 'solid-js/web'
-import { Routes } from 'generouted/solid-router'
+import { Routes } from '@generouted/solid-router'
 
 render(Routes, document.getElementById('app') as HTMLElement)
 ```
