+++ plugins/solid-router/src/plugin/generate.ts
