+++ plugins/solid-router/src/plugin/options.ts
