+++ plugins/solid-router/src/plugin/template.ts
